overlay: Add ruby_2_7_5 and ruby_2_7_4 as examples
This relies on the upstream builder, which does not provide exactly
great ergonomics for overrids.

We have to use the tagged build, rather than the released tarball. This
is because the default build relies on it to correcly configure
`useRailsExpress`. Turning this off would allow relying on the released
tarballs.

Building other minor releases (e.g. 2.6.*) will not work as overriding
`2_7` brings in patches that are unlikely to apply.

For this, relying on a local copy of the builders will be necessary.