modules: Drop usage of `nixpkgs` upstream module
This does not address the root cause of #6, the root cause of #6 was the
usage of `_module.args.pkgs`, which conflicts with the updated upstream
nixpkgs module.

Instead, we're *dropping* the module. Why? Because I believe it might
cause more issues than solve issues here.

In practice it's not needed for the use case of Fly, since all
evaluations start with a pre-defined `pkgs` parameter from the overlay.
In addition, it is not expected that users would change the reference of
Nixpkgs using that module.

Fixes #6.