modules/runtimes: Add ruby runtime config
Currently adds the Nixpkgs-native rubies, and the two sample
Fly-specific overlaid rubies as valid inputs.

This implementation provides "typedly" useful errors:

```
error: A definition for option `runtimes.ruby.version' is not of type
`one of "2.7.4", "2.7.5", "3.0.3", "3.1.0"'.
```

This was gotten when trying to use a Ruby version not declared as
available.

Work could be done to provide a custom specialization on the type where
the error message is made more useful for runtime selection.

Fixes https://github.com/fly-apps/nix-base/issues/12