runtimes/ruby: Add jemalloc option
"Fixes" https://github.com/fly-apps/nix-base/issues/1

The option had to be disabled for now, so it's kinda fixed.

The current implementation in upstream Nixpkgs does not forward the
jemalloc build input to gems being built. This means that native gems
will error with a message similar to:

```
/nix/store/z120avd72957v4dvr7azv1bgnzckhpmw-ruby-3.1.1/include/ruby-3.1.0/x86_64-linux/ruby/config.h:74:40: fatal error: jemalloc/jemalloc.h: No such file or directory
   74 | #define RUBY_ALTERNATIVE_MALLOC_HEADER <jemalloc/jemalloc.h>
```

This is entirely an upstream Nixpkgs packaging issue.